Embed a video into an HTML document

These functions are used to embed video into your rmarkdown html-documents, or into your shiny apps. There are functions to embed from YouTube, Vimeo, and Microsoft Channel 9 (who host the UseR! 2016 videos).

embed_vimeo(id, width = 500, height = 281, frameborder = 0,
  allowfullscreen = TRUE, query = NULL, fragment = NULL)

embed_youtube(id, width = 420, height = 315, frameborder = 0,
  allowfullscreen = TRUE, query = NULL)

embed_user2016(id, width = 560, height = 315, frameborder = 0,
  allowfullscreen = TRUE)

embed_user2017(id, width = 560, height = 315, frameborder = 0,
  allowfullscreen = TRUE)

embed_channel9(id, width = 560, height = 315, frameborder = 0,
  allowfullscreen = TRUE)

Arguments

id
character, identifier provided by the service

width
numeric, width of iframe (px)

height
numeric, height of iframe (px)

frameborder
numeric, size of frame border (px)

allowfullscreen
logical, indicates if to allow fullscreen

query
list of items to include in url-query string

fragment
character, string to include as url-fragment

Value

An embed object that prints an htmltools::tags$iframe element

Details

These services allow you to customize a lot of things by specifying an optional query string. The specification for the query string will differ according to the service being used:

YouTube

https://developers.google.com/youtube/player_parameters

Vimeo

https://developer.vimeo.com/player/embedding

Channel 9

https://channel9.msdn.com/Events/useR-international-R-User-conference/useR2016/Forty-years-of-S (click the Embed tab)

See also

use_start_time

Examples

embed_youtube("dQw4w9WgXcQ")

#> <iframe src="https://www.youtube.com/embed/dQw4w9WgXcQ" width="420" height="315" frameborder="0" allowfullscreen=""></iframe>

embed_vimeo("45196609")

#> <iframe class="vimeo-embed" src="https://player.vimeo.com/video/45196609" width="500" height="281" frameborder="0" webkitallowfullscreen="" mozallowfullscreen="" allowfullscreen=""></iframe>

embed_user2016("Literate-Programming")

#> <iframe src="https://channel9.msdn.com/Events/useR-international-R-User-conference/useR2016/Literate-Programming/player" width="560" height="315" frameborder="0" allowfullscreen=""></iframe>

embed_user2017("Room-202-Lightning-Talks") %>% use_start_time("26m35s")

#> <iframe src="https://channel9.msdn.com/Events/useR-international-R-User-conferences/useR-International-R-User-2017-Conference/Room-202-Lightning-Talks/player#time=0h26m35s:paused" width="560" height="315" frameborder="0" allowfullscreen=""></iframe>