Embed a video based on URL

You can use this function to embed video using only the URL and you do not need any customization beyond the start-time. It should work for all the services supported by the embed family of functions.

embed_url(url)

Arguments

url
character, URL of webpage for video

Value

An embed object that prints an htmltools::tags$iframe element

Details

This function calls suggest_embed then parses and evaluates the code. If you need to customize the iframe, suggest_embed may be more useful to you.

See also

suggest_embed

Examples

embed_url("https://youtu.be/1-vcErOPofQ?t=28s")

#> <iframe src="https://www.youtube.com/embed/1-vcErOPofQ?start=28" width="420" height="315" frameborder="0" allowfullscreen=""></iframe>