Specify a start time for an embedded video

This function provides you a consistent way to specify the start time, regardless of the service.

use_start_time(...)

# S3 method for default
use_start_time(...)

# S3 method for embed_youtube
use_start_time(embed, start_time, ...)

# S3 method for embed_vimeo
use_start_time(embed, start_time, ...)

# S3 method for embed_channel9
use_start_time(embed, start_time, is_paused = TRUE,
  ...)

Arguments

...

generic argumentss to pass through

embed

embed object made using an embed function

start_time

numeric (seconds), or character ("3m15s")

is_paused

logical, for "Channel 9" specifies if the video should be paused at this time

Details

The start_time argument can take a variety of formats; these inputs all evaluate to the same value:

"0h1m0s", "0h01m00s", "0h1m"

"1m0s", "1m"

"60s", 60

Please note that for Vimeo, you can specify a start time, but you can not specify that the video be paused at this time. In other words, it is like "autoplay" is set to TRUE, and you cannot unset it.

Examples

  rickroll_youtube() %>%
    use_start_time("3m32s")

#> <iframe src="https://www.youtube.com/embed/DLzxrzFCyOs?start=212" width="420" height="315" frameborder="0" allowfullscreen=""></iframe>