Embed video from service

These functions are used to embed video into your rmarkdown html-documents, or into your shiny apps. There are functions to embed from YouTube, Vimeo, Microsoft Channel 9 (who host the UseR! 2016 videos), and Box.

embed_user2016(
  id,
  width = NULL,
  height = 300,
  ratio = c("16by9", "4by3"),
  frameborder = 0,
  allowfullscreen = TRUE
)

embed_user2017(
  id,
  width = NULL,
  height = 300,
  ratio = c("16by9", "4by3"),
  frameborder = 0,
  allowfullscreen = TRUE
)

embed_box(
  id,
  custom_domain = getOption("vembedr.box_custom_domain"),
  width = NULL,
  height = 300,
  ratio = c("16by9", "4by3"),
  frameborder = 0,
  allowfullscreen = TRUE
)

embed_channel9(
  id,
  width = NULL,
  height = 300,
  ratio = c("16by9", "4by3"),
  frameborder = 0,
  allowfullscreen = TRUE
)

embed_vimeo(
  id,
  width = NULL,
  height = 300,
  ratio = c("16by9", "4by3"),
  frameborder = 0,
  allowfullscreen = TRUE,
  query = NULL,
  fragment = NULL
)

embed_youtube(
  id,
  width = NULL,
  height = 300,
  ratio = c("16by9", "4by3"),
  frameborder = 0,
  allowfullscreen = TRUE,
  query = NULL
)

Arguments

id
character, identifier provided by the service

width
numeric, width of iframe (px)

height
numeric, height of iframe (px)

ratio
character, indicates aspect ratio for the <iframe/>

frameborder
numeric, size of frame border (px)

allowfullscreen
logical, indicates if to allow fullscreen

custom_domain
character, (used by Box) name of Box-instance to use. It can be useful to use getOption("vembedr.box_custom_domain") if you are using a corporate instance of Box. If NULL, it will use the standard Box instance.

query
list, items to include in url-query string

fragment
character, string to include as url-fragment

Value

Object with S3 class vembedr_embed

Details

These services allow you to customize a lot of things by specifying an optional query string. The specification for the query string will differ according to the service being used:

YouTube

https://developers.google.com/youtube/player_parameters

Vimeo

https://developer.vimeo.com/player/embedding

Channel 9

https://channel9.msdn.com/Events/useR-international-R-User-conference/useR2016/Forty-years-of-S (click the Embed tab)

Box

https://developer.box.com/docs/box-embed#section-build-box-embed-programatically

See also

Examples

embed_youtube("dQw4w9WgXcQ")

#> <div class="vembedr">
#>   <div>
#>     <iframe src="https://www.youtube.com/embed/dQw4w9WgXcQ" width="533" height="300" frameborder="0" allowfullscreen=""></iframe>
#>   </div>
#> </div>

embed_vimeo("45196609")

#> <div class="vembedr">
#>   <div>
#>     <iframe class="vimeo-embed" src="https://player.vimeo.com/video/45196609" width="533" height="300" frameborder="0" webkitallowfullscreen="" mozallowfullscreen="" allowfullscreen=""></iframe>
#>   </div>
#> </div>

embed_user2016("Literate-Programming")

#> <div class="vembedr">
#>   <div>
#>     <iframe src="https://channel9.msdn.com/Events/useR-international-R-User-conference/useR2016/Literate-Programming/player" width="533" height="300" frameborder="0" allowfullscreen=""></iframe>
#>   </div>
#> </div>

embed_user2017("Room-202-Lightning-Talks") %>% use_start_time("26m35s")

#> <div class="vembedr">
#>   <div>
#>     <iframe src="https://channel9.msdn.com/Events/useR-international-R-User-conferences/useR-International-R-User-2017-Conference/Room-202-Lightning-Talks/player#time=0h26m35s:paused" width="533" height="300" frameborder="0" allowfullscreen=""></iframe>
#>   </div>
#> </div>

embed_box("m5do45hvzw32iv2aors3urf5pgkxxazx")

#> <div class="vembedr">
#>   <div>
#>     <iframe src="https://app.box.com/embed/s/m5do45hvzw32iv2aors3urf5pgkxxazx" width="533" height="360" frameborder="0" allowfullscreen="" webkitallowfullscreen="" msallowfullscreen=""></iframe>
#>   </div>
#> </div>