Embed a popular video

If you want to experiment with the arguments to embed(), such as query, but do not have a particular video in mind, this function may be useful to you.

rickroll_vimeo(...)

rickroll_youtube(...)

rickroll_channel9(...)

Arguments

...
arguments (other than id) passed on to embed()

Value

An embed object that prints an htmltools::tags$iframe element

Examples

  rickroll_vimeo()

#> <div>
#>   <div>
#>     <iframe class="vimeo-embed" src="https://player.vimeo.com/video/148751763" width="500" height="281" frameborder="0" webkitallowfullscreen="" mozallowfullscreen="" allowfullscreen=""></iframe>
#>   </div>
#> </div>

  rickroll_youtube()

#> <div>
#>   <div>
#>     <iframe src="https://www.youtube.com/embed/DLzxrzFCyOs" width="420" height="315" frameborder="0" allowfullscreen=""></iframe>
#>   </div>
#> </div>

  rickroll_channel9()

#> <div>
#>   <div>
#>     <iframe src="https://channel9.msdn.com/Blogs/Dan/BlueHat-v7-Katie-Moussouris-interviews-Dan-Kaminsky-on-some-interesting-research-hes-been-doing-late/player#time=0h5m8s:paused" width="560" height="315" frameborder="0" allowfullscreen=""></iframe>
#>   </div>
#> </div>